+++ libs/react/src/components/molecules/editEntityAssociationsModal/editEntityAssociationsModal.tsx
@@ -6,7 +6,7 @@ import { flowbiteThemeOverride } from '@coldpbc/themes';
 import {useAddToastMessage, useAuth0Wrapper, useColdContext, useEntityData, useUpdateEntityAssociations} from '@coldpbc/hooks';
 import {GridCellParams, GridColDef, GridToolbarContainer, GridToolbarQuickFilter} from '@mui/x-data-grid';
 import { Checkbox } from '@mui/material';
-import {isEqual, sortBy} from 'lodash';
+import { isEqual, lowerCase, sortBy } from 'lodash';
 import { FetchResult } from '@apollo/client';
 import { ToastMessage } from '@coldpbc/interfaces';
 
@@ -14,7 +14,8 @@ interface EditEntityAssociationsModalProps {
 	buttonText: string;
 	refresh: () => void;
 	title: string;
-	entityLevel: EntityLevel;
+	entityLevelToAdd: EntityLevel;
+  entityLevelToBeAddedTo: EntityLevel;
 	idsSelected: string[];
 	saveButtonText: string;
 	entityToBeAddedId: string;
@@ -22,13 +23,13 @@ interface EditEntityAssociationsModalProps {
 }
 
 export const EditEntityAssociationsModal = (
-  {buttonText, refresh, title, entityLevel, idsSelected, entityToBeAddedId}: EditEntityAssociationsModalProps
+  {buttonText, refresh, title, entityLevelToAdd, entityLevelToBeAddedTo, idsSelected, entityToBeAddedId}: EditEntityAssociationsModalProps
 ) => {
   const {orgId} = useAuth0Wrapper();
   const [showEntityAssociationModal, setShowEntityAssociationModal] = useState<boolean>(false);
   const [rowsSelected, setRowsSelected] = useState<string[]>([]);
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const entities = useEntityData(entityLevel === EntityLevel.ORGANIZATION ? undefined : entityLevel, orgId);
+  const entities = useEntityData(entityLevelToAdd === EntityLevel.ORGANIZATION ? undefined : entityLevelToAdd, orgId);
   const {addToastMessage} = useAddToastMessage();
   const {logBrowser} = useColdContext();
   const {callMutateFunction} = useUpdateEntityAssociations();
@@ -91,59 +92,47 @@ export const EditEntityAssociationsModal = (
     },
   ];
 
-  const onEntitiesUpdate = () => {
+  const onEntitiesUpdate = async () => {
     setIsLoading(true);
     const removedRows = idsSelected.filter(id => !rowsSelected.includes(id));
     const addedRows = rowsSelected.filter(id => !idsSelected.includes(id));
     try {
       const promises: (Promise<void> | Promise<FetchResult<any>>)[] = []
       addedRows.forEach(row => {
-        promises.push(callMutateFunction(entityLevel, row, entityToBeAddedId, orgId, true))
+        promises.push(callMutateFunction(entityLevelToAdd, entityLevelToBeAddedTo, row, entityToBeAddedId, orgId, 'add'))
       })
       removedRows.forEach(row => {
-        promises.push(callMutateFunction(entityLevel, row, entityToBeAddedId, orgId, false))
+        promises.push(callMutateFunction(entityLevelToAdd, entityLevelToBeAddedTo, row, entityToBeAddedId, orgId, 'delete'))
       });
 
-      Promise.all(promises).then((responses) => {
-        logBrowser(`Updated entity associations successfully`, 'info', {
-          orgId,
-          entityLevel,
-          entityToBeAddedId,
-          idsSelected,
-          rowsSelected,
-          responses
-        });
-        addToastMessage({
-          message: 'Associations updated successfully',
-          type: ToastMessage.SUCCESS
-        })
-        refresh();
-        setShowEntityAssociationModal(false);
-      }).catch((error) => {
-        logBrowser(`Error updating entity associations`, 'error', {
-          orgId,
-          entityLevel,
-          entityToBeAddedId,
-          idsSelected,
-          rowsSelected,
-          error
-        }, error)
-        addToastMessage({
-          message: 'Error updating associations',
-          type: ToastMessage.FAILURE
-        })
-      })
-    } catch (error) {
-      logBrowser(`Error updating entity associations`, 'error', {
+      const responses = await Promise.all(promises)
+      logBrowser(`Updated ${lowerCase(entityLevelToAdd)}s successfully`, 'info', {
         orgId,
-        entityLevel,
+        entityLevelToAdd,
+        entityLevelToBeAddedTo,
+        entityToBeAddedId,
+        idsSelected,
+        rowsSelected,
+        responses
+      });
+      addToastMessage({
+        message: `Updated ${lowerCase(entityLevelToBeAddedTo)} successfully`,
+        type: ToastMessage.SUCCESS
+      })
+      refresh();
+      setShowEntityAssociationModal(false);
+    } catch (error) {
+      logBrowser(`Error updating ${lowerCase(entityLevelToBeAddedTo)}`, 'error', {
+        orgId,
+        entityLevelToAdd,
+        entityLevelToBeAddedTo,
         entityToBeAddedId,
         idsSelected,
         rowsSelected,
         error
       }, error)
       addToastMessage({
-        message: 'Error updating associations',
+        message: `Error updating ${lowerCase(entityLevelToBeAddedTo)}`,
         type: ToastMessage.FAILURE
       })
     } finally {
@@ -185,7 +174,11 @@ export const EditEntityAssociationsModal = (
 						</div>
 					</div>
 					<div className="w-full flex flex-row justify-between">
-						<BaseButton label="Cancel" onClick={() => setShowEntityAssociationModal(false)} variant={ButtonTypes.secondary} disabled={isLoading} />
+						<BaseButton
+              label="Cancel"
+              onClick={() => setShowEntityAssociationModal(false)}
+              variant={ButtonTypes.secondary}
+            />
 						<div className="flex flex-row gap-[16px] items-center">
 							<div className="text-body font-bold text-tc-secondary">
 								{rowsSelected.length}/{rows.length} Selected