+++ libs/react/src/lib/graphql/mutateQueries.ts
@@ -183,3 +183,17 @@ export const CREATE_PRODUCT = gql`
     }
   }
 `;
+
+export const UPDATE_PRODUCT_MATERIAL = gql`
+  mutation UpdateProductMaterial($input: ProductMaterialUpdateInput!) {
+    updateProductMaterial(input: $input) {
+      id
+    }
+  }
+`;
+
+export const DELETE_PRODUCT_MATERIALS = gql`
+  mutation DeleteProductMaterials($filter: ProductMaterialsListFilter!) {
+    deleteProductMaterials(filter: $filter)
+  }
+`;